# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Документация для RequestController",
        "description": "Контроллер для управления заявками (requests): создание, получение списка, обновление.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/requests": {
            "get": {
                "tags": [
                    "Request"
                ],
                "summary": "Получить список заявок",
                "description": "Позволяет получить список заявок с возможностью фильтрации через параметры запроса.",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу заявки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Active"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Request"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса"
                    }
                }
            },
            "put": {
                "tags": [
                    "Request"
                ],
                "summary": "Обновить заявку",
                "description": "Обновляет заявку и отправляет email-уведомление о завершении.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID заявки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Новое имя заявителя",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Новый email",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Новый статус заявки",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Resolved"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявка успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Request"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "post": {
                "tags": [
                    "Request"
                ],
                "summary": "Создать новую заявку",
                "description": "Создаёт новую заявку на основе переданных данных.",
                "operationId": "save",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя заявителя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email заявителя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Сообщение",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Заявка успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Request"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Request": {
                "title": "Request",
                "description": "Модель заявки",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "$message": {
                        "type": "string",
                        "example": "Тестовое сообщение"
                    },
                    "status": {
                        "description": "Статус заявки: Active, Resolved",
                        "type": "string",
                        "example": "Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Request",
            "description": "Операции с заявками пользователей"
        }
    ]
}